REMOVE PRE-COMMIT PRETTIER HOOK
Remove the pre-commit prettier configuration from package.json as it is causing issues. Prettier formatting will still happen during the build process.

The pre-commit hook was configuring prettier to run on each commit, formatting files. This was causing issues with commits. The removal of this configuration means prettier will still format files when the code is built, but will no longer block commits.

Write commit messages in the imperative tense.